azure-functions/daily-search/requirements.txt
azure-functions
httpx>=0.27
